+++ src/main/java/com/zhiend/projectms/entity/DevelopmentStatus.java
@@ -33,12 +33,12 @@ public class DevelopmentStatus implements Serializable {
      * 主键，自增
      */
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 项目ID
      */
-    private Integer projectId;
+    private Long projectId;
 
     /**
      * 当前状态，改为枚举类型
