+++ src/main/java/com/zhiend/projectms/entity/RiskTracking.java
@@ -28,12 +28,12 @@ public class RiskTracking implements Serializable {
      * 主键，自增
      */
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 项目ID
      */
-    private Integer projectId;
+    private Long projectId;
 
     /**
      * 风险属性
